+++ Assets/Scripts/TerrainGen/Blocks.cs
@@ -2,14 +2,14 @@ using Unity.Mathematics;
 
 public class Blocks
 {
-	public static readonly int Air = 0;
-	public static readonly int Stone = 1;
-	public static readonly int Dirt = 2;
-	public static readonly int Grass = 3;
-	public static readonly int Sand = 4;
-	public static readonly int Water = 5;
-	public static readonly int Log = 6;
-	public static readonly int Leaves = 7;
+	public static readonly byte Air = 0;
+	public static readonly byte Stone = 1;
+	public static readonly byte Dirt = 2;
+	public static readonly byte Grass = 3;
+	public static readonly byte Sand = 4;
+	public static readonly byte Water = 5;
+	public static readonly byte Log = 6;
+	public static readonly byte Leaves = 7;
 
 	//The bottom left UVs for all 6 faces
 	public static FaceUVs GetUVs(int blockIndex)
